Revert "Move Platform StrEnum to const (#60857)"
This reverts commit 6e220d5d17c828dc6705fc23cbde4dc7230b3226.
