+++ homeassistant/components/wled/__init__.py
@@ -2,8 +2,8 @@
 from __future__ import annotations
 
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers.entity_platform import Platform
 
 from .const import DOMAIN
 from .coordinator import WLEDDataUpdateCoordinator
