+++ homeassistant/util/dt.py
@@ -10,6 +10,8 @@ from typing import Any, cast
 
 import ciso8601
 
+from homeassistant.const import MATCH_ALL
+
 if sys.version_info[:2] >= (3, 9):
     import zoneinfo
 else:
@@ -213,7 +215,7 @@ def get_age(date: dt.datetime) -> str:
 
 def parse_time_expression(parameter: Any, min_value: int, max_value: int) -> list[int]:
     """Parse the time expression part and return a list of times to match."""
-    if parameter is None or parameter == "*":
+    if parameter is None or parameter == MATCH_ALL:
         res = list(range(min_value, max_value + 1))
     elif isinstance(parameter, str):
         if parameter.startswith("/"):
