+++ script/hassfest/dependencies.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 import ast
 from pathlib import Path
 
-from homeassistant.const import Platform
 from homeassistant.requirements import DISCOVERY_INTEGRATIONS
+from homeassistant.setup import BASE_PLATFORMS
 
 from .model import Integration
 
@@ -91,11 +91,12 @@ class ImportCollector(ast.NodeVisitor):
 
 
 ALLOWED_USED_COMPONENTS = {
-    *{platform.value for platform in Platform},
     # Internal integrations
     "alert",
     "automation",
+    "button",
     "conversation",
+    "button",
     "device_automation",
     "frontend",
     "group",
@@ -118,6 +119,8 @@ ALLOWED_USED_COMPONENTS = {
     "webhook",
     "websocket_api",
     "zone",
+    # Entity integrations with platforms
+    *BASE_PLATFORMS,
     # Other
     "mjpeg",  # base class, has no reqs or component to load.
     "stream",  # Stream cannot install on all systems, can be imported without reqs.
